# internal/controlapi/models.go
// Package controlapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package controlapi

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ErrorCode.
const (
	Conflict              ErrorCode = "conflict"
	DependencyUnavailable ErrorCode = "dependency_unavailable"
	Internal              ErrorCode = "internal"
	InvalidArgument       ErrorCode = "invalid_argument"
	NotFound              ErrorCode = "not_found"
	PermissionDenied      ErrorCode = "permission_denied"
	RateLimited           ErrorCode = "rate_limited"
	Unauthenticated       ErrorCode = "unauthenticated"
)

// Defines values for SessionState.
const (
	Active     SessionState = "active"
	Terminated SessionState = "terminated"
)

// Defines values for TerminationReason.
const (
	AdminTerminated TerminationReason = "admin_terminated"
	Error           TerminationReason = "error"
	MaxDuration     TerminationReason = "max_duration"
	Revoked         TerminationReason = "revoked"
)

// Defines values for OrderBy.
const (
	OrderByStartTimeAsc  OrderBy = "start_time asc"
	OrderByStartTimeDesc OrderBy = "start_time desc"
)

// Defines values for ListSessionsParamsOrderBy.
const (
	ListSessionsParamsOrderByStartTimeAsc  ListSessionsParamsOrderBy = "start_time asc"
	ListSessionsParamsOrderByStartTimeDesc ListSessionsParamsOrderBy = "start_time desc"
)

// ErrorCode Stable error codes.
type ErrorCode string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Code Stable error codes.
	Code ErrorCode `json:"code"`

	// Details Optional structured error details.
	Details *map[string]interface{} `json:"details,omitempty"`

	// Message Human-readable error message.
	Message string `json:"message"`

	// RequestId Server-generated request identifier.
	RequestId string `json:"request_id"`
}

// Session defines model for Session.
type Session struct {
	// EndTime Session end time (UTC); omitted while active.
	EndTime *time.Time `json:"end_time,omitempty"`

	// PublicKeyFingerprint SSH public key fingerprint (SHA256).
	PublicKeyFingerprint string `json:"public_key_fingerprint"`

	// SessionId Session identifier.
	SessionId openapi_types.UUID `json:"session_id"`

	// SourceIp Source IP address (IPv4 or IPv6).
	SourceIp string `json:"source_ip"`

	// StartTime Session start time (UTC).
	StartTime time.Time `json:"start_time"`

	// State Session lifecycle state.
	State SessionState `json:"state"`

	// TerminationReason Reason a session was terminated.
	TerminationReason *TerminationReason `json:"termination_reason,omitempty"`

	// UserName User identity derived from the User CRD.
	UserName string `json:"user_name"`
}

// SessionListResponse defines model for SessionListResponse.
type SessionListResponse struct {
	Sessions []Session `json:"sessions"`
}

// SessionState Session lifecycle state.
type SessionState string

// TerminationReason Reason a session was terminated.
type TerminationReason string

// Limit defines model for limit.
type Limit = int

// Offset defines model for offset.
type Offset = int

// OrderBy defines model for order_by.
type OrderBy string

// SessionId defines model for session_id.
type SessionId = openapi_types.UUID

// ErrorDependencyUnavailable defines model for ErrorDependencyUnavailable.
type ErrorDependencyUnavailable = ErrorResponse

// ErrorInternal defines model for ErrorInternal.
type ErrorInternal = ErrorResponse

// ErrorInvalidArgument defines model for ErrorInvalidArgument.
type ErrorInvalidArgument = ErrorResponse

// ErrorNotFound defines model for ErrorNotFound.
type ErrorNotFound = ErrorResponse

// ErrorPermissionDenied defines model for ErrorPermissionDenied.
type ErrorPermissionDenied = ErrorResponse

// ErrorRateLimited defines model for ErrorRateLimited.
type ErrorRateLimited = ErrorResponse

// ErrorUnauthenticated defines model for ErrorUnauthenticated.
type ErrorUnauthenticated = ErrorResponse

// ListSessionsParams defines parameters for ListSessions.
type ListSessionsParams struct {
	// Limit Maximum number of sessions to return. Defaults to 50 and is capped at 200.
	Limit *Limit `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Number of sessions to skip before returning results. Defaults to 0.
	Offset *Offset `form:"offset,omitempty" json:"offset,omitempty"`

	// OrderBy Sort order. Default: start_time desc.
	OrderBy *ListSessionsParamsOrderBy `form:"order_by,omitempty" json:"order_by,omitempty"`
}

// ListSessionsParamsOrderBy defines parameters for ListSessions.
type ListSessionsParamsOrderBy string